refactor actionFurtherAction classes into a separate package for the sake of clarity
