+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/actionfurtherevidence/FurtherEvidenceActionDynamicListItems.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.ccd.presubmit;
+package uk.gov.hmcts.reform.sscs.ccd.presubmit.actionfurtherevidence;
 
 import lombok.Getter;
